+++ src/rubysys/types.rs
@@ -17,6 +17,13 @@ pub struct EncodingIndex(pub c_int);
 pub type Argc = c_int;
 pub type CallbackPtr = *const c_void;
 pub type CallbackMutPtr = *mut c_void;
+pub type BlockCallFunction = extern fn(
+    yielded_arg: Value,
+    callback_arg: Value,
+    argc: c_int,
+    argv: *const Value,
+    block_arg: Value,
+) -> Value;
 
 #[repr(C)]
 pub struct RBasic {